+++ changelog.md
@@ -5,7 +5,7 @@
 - **Fixed:**
   - Character manager: Fix context menu items not working
 - **Changed:**
-  - Update better-sqlite3 7.5.3 to 7.6.0
+  - Update better-sqlite3 7.5.3 to 7.6.1
   - Update moment 2.29.3 to 2.29.4
   - Update electron 19.0.7 to 19.0.8